// src/data.types.ts
export interface PokemonData {
  count: number;
  next: string | null;
  previous: string | null;
  name: string;
  results: PokemonUrl[];
}

export interface PokemonUrl {
  url: string;
  name: string;
}

export interface Pokemon {
  abilities: Ability[];
  base_experience: number;
  forms: Species[];
  game_indices: GameIndex[];
  height: number;
  held_items: any[];
  id: number;
  is_default: boolean;
  location_area_encounters: string;
  moves: Move[];
  name: string;
  order: number;
  past_types: any[];
  species: Species;
  sprites: Sprites;
  stats: Stat[];
  types: Type[];
  weight: number;
}

export interface PokemonType {
  damage_relations: DamageRelations;
  game_indices: GameGeneration[];
  generation: Species;
  id: number;
  move_damage_class: Species;
  moves: Species[];
  name: string;
  names: Name[];
  past_damage_relations: any[];
  pokemon: PokemonTypeUrl[];
}

export interface PokemonTypeUrl {
  pokemon: PokemonUrl;
}

export interface Name {
  language: Species;
  name: string;
}

export interface GameGeneration {
  game_index: number;
  generation: Species;
}

export interface Ability {
  ability: Species;
  is_hidden: boolean;
  slot: number;
}

export interface Species {
  name: string;
  url: string;
}

export interface GameIndex {
  game_index: number;
  version: Species;
}

export interface Move {
  move: Species;
  version_group_details: VersionGroupDetail[];
}

export interface VersionGroupDetail {
  level_learned_at: number;
  move_learn_method: Species;
  version_group: Species;
}

export interface GenerationV {
  "black-white": Sprites;
}

export interface GenerationIv {
  "diamond-pearl": Sprites;
  "heartgold-soulsilver": Sprites;
  platinum: Sprites;
}

export interface Versions {
  "generation-i": GenerationI;
  "generation-ii": GenerationIi;
  "generation-iii": GenerationIii;
  "generation-iv": GenerationIv;
  "generation-v": GenerationV;
  "generation-vi": { [key: string]: Home };
  "generation-vii": GenerationVii;
  "generation-viii": GenerationViii;
}

export interface Sprites {
  back_default: string;
  back_female: string;
  back_shiny: string;
  back_shiny_female: string;
  front_default: string;
  front_female: string;
  front_shiny: string;
  front_shiny_female: string;
  other?: Other;
  versions?: Versions;
  animated?: Sprites;
}
// ...
export interface Other {
  dream_world: DreamWorld;
  home: Home;
  "official-artwork": OfficialArtwork;
}

export interface Stat {
  base_stat: number;
  effort: number;
  stat: Species;
}

export interface Type {
  slot: number;
  type: Species;
}

export interface DamageRelations {
  double_damage_from: Species[];
  double_damage_to: Species[];
  half_damage_from: Species[];
  half_damage_to: Species[];
  no_damage_from: Species[];
  no_damage_to: Species[];
}

export interface PokemonTypeData {
  typeName: string;
  typeRelation: DamageRelations;
  pokemons: PokemonTypeUrl[];
}